divide qc file, rm redudant functions
